implement 'count_pips()' helper function
In `count_pips()` I count the frequency of each pip in a 7-card poker
hand and update the frequency count in an int array of size 14.

I also fixed a logic error in `main()` in which I wasn't creating a new
deck and shuffling it within the 10-iteration loop.

**NEXT STEP**: create a helper function to _classify_ each hand, i.e.
"Four of a Kind", "Full House", "Three of a Kind", "Two Pair", etc.